web/requirements.txt
Django==1.11
django-appconf==1.0.2	
django-bootstrap-breadcrumbs==0.9.0
django-ckeditor==5.3.1
django-classy-tags==0.8.0
django-contact-form==1.5
django-dbbackup==3.2.0
django-disqus==0.5
django-haystack==2.8.1
django-recaptcha==1.3.1
django-storages==1.6.5
django-twitter-tag==1.2.1
redis==2.10.6
django-redis==4.8.0
gunicorn==19.7.1
boto==2.48.0
dropbox==8.5.0
feedparser==5.2.1
funcsigs==1.0.2
mock==2.0.0
oauthlib==2.0.6
pbr==3.1.1
Pillow==2.7.0
psycopg2==2.7.3
pytz==2017.3
Pyzotero==1.3.0
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tablib==0.12.1
twitter==1.18.0
urllib3==1.22
Whoosh==2.7.4

